feat(profile): profiles are now identified via unique IDs
Profiles now both have an ID attribute and a Name attribute. The ID uniquely identifies the profile against others and is generated at random upon generation, and the name is a changable identifier for the end user. This introduces a breaking change, as profile and annotation files have to be updated accordingly.