feat(text-model): annotation classes are now identified via unique IDs
Annotation classes now both have an ID attribute and a Name attribute. The ID uniquely identifies the annotation class against others and is generated at random upon generation, and the name is a changable identifier for the end user. This introduces a breaking change, as profile and annotation files have to be updated accordingly.